Implementation/Windows/commons/requirements.txt
Flask == 2.1.3
FMPy == 0.3.9
kafka-python == 2.0.2
OMSimulator == 2.1.1